+++ CHANGELOG.md
@@ -4,10 +4,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/0.3.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+
+## [1.1.0] - 2017-04-19
 ### Added
 - max vote, previous default was 100 now it's 10. Bump DB from 2 to 3
+- alt plus: increase font
+- alt minus: decrease font
+- auto send votes to display
+
 ### Fixed
+- report: dump all users
+- ui: support from 0 to nUsers included
+
 ### Changed
+- report trials: replace min and max values with '-'
+- increase trials limit to 10
 
 ## [1.0.1] - 2016-04-24
 ### First public release